altera tipos para permitir o uso do mysql

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -27,7 +27,7 @@ class CreateUsersTable extends Migration
             $table->rememberToken();
             $table->timestamps();
 
-            $table->integer('enderecoId')->nullable();
+            $table->unsignedBigInteger('enderecoId')->nullable();
         });
     }
