+++ database/migrations/2020_02_05_123101_create_recursos_table.php
@@ -20,8 +20,8 @@ class CreateRecursosTable extends Migration
             $table->text('corpoRecurso');
             $table->boolean('statusAvaliacao');
 
-            $table->integer('trabalhoId');
-            $table->integer('comissaoId');
+            $table->unsignedBigInteger('trabalhoId');
+            $table->unsignedBigInteger('comissaoId');
         });
     }
