+++ database/migrations/2020_02_05_123139_create_areas_table.php
@@ -18,8 +18,8 @@ class CreateAreasTable extends Migration
             $table->timestamps();
             $table->string('nome');
 
-            // $table->integer('modalidadeId')->nullable();
-            $table->integer('eventoId');
+            // $table->unsignedBigInteger('modalidadeId')->nullable();
+            $table->unsignedBigInteger('eventoId');
         });
     }
