+++ database/migrations/2021_08_20_090208_create_arquivo_correcaos_table.php
@@ -16,7 +16,7 @@ class CreateArquivoCorrecaosTable extends Migration
         Schema::create('arquivo_correcaos', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('caminho');
-            $table->integer('trabalhoId');
+            $table->unsignedBigInteger('trabalhoId');
             $table->timestamps();
         });
     }
